Relax requirement for objects v2 membership get UUID param
The docs indicate that the UUID being omited as a param when calling
getMemberships on objects v2 should use the client UUID. The validation
was a bit too strict, so this relaxes that. We also update the code
to use the client UUID by default if the params UUID is  missing.