fix(shared-worker): fix rapid subscribe aggregate
Fix the issue because of which rapid subscription (from other PubNub clients in response to
received subscribe response) throttling causes delayed channel list change.

refactor(shared-worker): add output of requests processed by the `SharedWorker`

Modify the `log` payload for `SharedWorker` to make it possible to log sent / received requests
information to the main browser window (not to the SharedWorker console).