refactor(shared-worker): react only on actual 'auth' and 'userId' change
Adjust the code that is called on identity (`userId`) and auth token change to modify the state
only if values actually have been changed.

refactor(shared-worker): change condition for stalled `ping` timer

Change the condition that is used to identify whether the `offline` detection timer has been
suspended by the browser or not before trying to evict "offline" PubNub clients.